refactor: split dftu_pw into dftu_pw_tools (pure functions) + dftu_pw (orchestration)
- dftu_pw_tools.h/cpp: 7 pure functions in namespace pw, unit-testable
- dftu_pw.h/cpp: cal_occ_pw + accumulate_occ_one_k in namespace DFTU_BASE
- test_dftu_pw_tools.cpp: tests for pure functions, links only dftu_pw_tools.cpp
- Update CMakeLists.txt accordingly
