+++ source/source_lcao/module_dftu/test/CMakeLists.txt
@@ -1,13 +1,6 @@
 abacus_disable_feature_definitions(__CUDA)
 
 AddTest(
-  TARGET dftu_pw_test
-  LIBS base device parameter
-  SOURCES dftu_pw_test.cpp
-                ../../../source_pw/module_pwdft/dftu_base_tools.cpp
-)
-
-AddTest(
   TARGET dftu_core_test
   LIBS base device
   SOURCES dftu_core_test.cpp
