+++ source/source_lcao/module_dftu/test/dftu_operator_test.cpp
@@ -323,77 +323,3 @@ TEST_F(VoigtToMatrixTest, FullMappingAndSymmetry)
     EXPECT_NEAR(matrix[1], matrix[3], 1e-15); EXPECT_NEAR(matrix[2], matrix[6], 1e-15);
     EXPECT_NEAR(matrix[5], matrix[7], 1e-15);
 }
-
-// =====================================================================
-// 7. PW operator index setup (ip_iat, ip_m, pot_onsite_begin_iat)
-// ip_m[ip] = m index if projector is correlated, else -1
-// ip_iat[ip] = atom index, pot_onsite_begin_iat[iat] = pot_onsite array offset
-// =====================================================================
-
-class PWIndexSetupTest : public ::testing::Test
-{
-  protected:
-    struct AtomInfo { int it, nh, target_l; }; // target_l=-1 if not correlated
-
-    void setup_indices(const std::vector<AtomInfo>& atoms,
-        std::vector<int>& ip_iat, std::vector<int>& ip_m,
-        std::vector<int>& pot_onsite_begin_iat, int& pot_onsite_total_size)
-    {
-        int ip0 = 0, pot_onsite_begin = 0, npol = 1;
-        ip_iat.resize(0); ip_m.resize(0); pot_onsite_begin_iat.resize(atoms.size());
-        for (const auto& atom : atoms)
-        {
-            ip_iat.resize(ip_iat.size() + atom.nh);
-            ip_m.resize(ip_m.size() + atom.nh);
-            if (atom.target_l == -1)
-            {
-                for (int ip = 0; ip < atom.nh; ip++)
-                { ip_iat[ip0] = static_cast<int>(&atom - &atoms[0]); ip_m[ip0++] = -1; }
-                pot_onsite_begin_iat[&atom - &atoms[0]] = 0;
-            }
-            else
-            {
-                int tlp1 = 2 * atom.target_l + 1;
-                pot_onsite_begin_iat[&atom - &atoms[0]] = pot_onsite_begin;
-                pot_onsite_begin += tlp1 * tlp1 * npol * npol;
-                int m_begin = atom.target_l * atom.target_l;
-                int m_end = (atom.target_l + 1) * (atom.target_l + 1);
-                for (int ip = 0; ip < atom.nh; ip++)
-                {
-                    ip_iat[ip0] = static_cast<int>(&atom - &atoms[0]);
-                    ip_m[ip0++] = (ip >= m_begin && ip < m_end) ? ip - m_begin : -1;
-                }
-            }
-        }
-        pot_onsite_total_size = pot_onsite_begin;
-    }
-};
-
-TEST_F(PWIndexSetupTest, SingleCorrelatedAtom_DOrbital)
-{
-    std::vector<AtomInfo> atoms = {{0, 9, 2}}; // s(1)+p(3)+d(5) projectors, l=2
-    std::vector<int> ip_iat, ip_m, pot_onsite_begin_iat; int pot_onsite_total_size;
-    setup_indices(atoms, ip_iat, ip_m, pot_onsite_begin_iat, pot_onsite_total_size);
-    // Projectors 0-3 (s+p): m=-1; 4-8 (d): m=0..4
-    EXPECT_EQ(ip_iat.size(), 9u);
-    for (int ip = 0; ip < 4; ip++) EXPECT_EQ(ip_m[ip], -1);
-    for (int ip = 4; ip < 9; ip++) { EXPECT_EQ(ip_iat[ip], 0); EXPECT_EQ(ip_m[ip], ip-4); }
-    EXPECT_EQ(pot_onsite_begin_iat[0], 0);
-    EXPECT_EQ(pot_onsite_total_size, 25); // 5*5
-}
-
-TEST_F(PWIndexSetupTest, MixedCorrelatedUncorrelated)
-{
-    std::vector<AtomInfo> atoms = {{0, 4, 1}, {1, 2, -1}}; // atom0: p-correlated, atom1: not
-    std::vector<int> ip_iat, ip_m, pot_onsite_begin_iat; int pot_onsite_total_size;
-    setup_indices(atoms, ip_iat, ip_m, pot_onsite_begin_iat, pot_onsite_total_size);
-    // atom0: s(ip=0)->m=-1, p(ip=1,2,3)->m=0,1,2
-    EXPECT_EQ(ip_iat[0], 0); EXPECT_EQ(ip_m[0], -1);
-    EXPECT_EQ(ip_iat[1], 0); EXPECT_EQ(ip_m[1], 0);
-    EXPECT_EQ(ip_iat[2], 0); EXPECT_EQ(ip_m[2], 1);
-    EXPECT_EQ(ip_iat[3], 0); EXPECT_EQ(ip_m[3], 2);
-    // atom1: all m=-1
-    EXPECT_EQ(ip_iat[4], 1); EXPECT_EQ(ip_m[4], -1);
-    EXPECT_EQ(ip_iat[5], 1); EXPECT_EQ(ip_m[5], -1);
-    EXPECT_EQ(pot_onsite_total_size, 9); // 3*3 for p-orbital
-}
