+++ source/source_pw/module_pwdft/CMakeLists.txt
@@ -14,8 +14,8 @@ list(APPEND objects
     op_pw_exx_pot.cpp
     dftu_base.cpp
     dftu_base_io.cpp
-    dftu_base_occ.cpp
-    dftu_base_tools.cpp
+    dftu_pw.cpp
+    dftu_pw_tools.cpp
     yukawa_screening.cpp
     setup_pot.cpp
     setup_pwrho.cpp
