+++ source/source_pw/module_pwdft/dftu_base.cpp
@@ -227,6 +227,6 @@ bool Plus_U_Base::u_converged()
 
 
 // cal_occ_pw() is implemented as free function DFTU_BASE::cal_occ_pw
-// in source_pw/module_pwdft/dftu_base_occ.cpp.
-// Pure per-atom kernels live in dftu_base_tools.{h,cpp}
+// in source_pw/module_pwdft/dftu_pw.cpp.
+// All pure per-atom kernels also live in dftu_pw.{h,cpp}
 // as free functions in namespace DFTU_BASE.
