+++ source/source_pw/module_pwdft/dftu_pw_tools.cpp
@@ -1,6 +1,11 @@
-#include "source_pw/module_pwdft/dftu_base_tools.h"
+#include "source_pw/module_pwdft/dftu_pw_tools.h"
 
-namespace DFTU_BASE {
+#include "source_pw/module_pwdft/dftu_base.h"
+#include "source_cell/unitcell.h"
+#include "source_base/parallel_reduce.h"
+#include "source_base/global_variable.h"
+
+namespace pw {
 
 void pauli_to_spin_basis(std::complex<double>* pot_onsite, int m_size)
 {
@@ -146,4 +151,105 @@ void accumulate_occ_scalar(
     }
 }
 
-} // namespace DFTU_BASE
+void reduce_occ_mat(const UnitCell& cell,
+                    const int nspin,
+                    const int kpar,
+                    const std::vector<int>& l_channel,
+                    OccupationMatrix& occmat)
+{
+    for(int iat = 0; iat < cell.nat; iat++)
+    {
+        const int it = cell.iat2it[iat];
+        const int target_l = l_channel[it];
+        if(target_l == -1)
+        {
+            continue;
+        }
+        const int size = (2 * target_l + 1) * (2 * target_l + 1);
+
+        if(nspin != 4)
+        {
+            Parallel_Reduce::reduce_double_allpool(kpar,
+                    GlobalV::NPROC_IN_POOL,
+                    occmat.mat(iat, target_l, 0, 0).c,
+                    size);
+            if(nspin == 2)
+            {
+                Parallel_Reduce::reduce_double_allpool(kpar,
+                        GlobalV::NPROC_IN_POOL,
+                        occmat.mat(iat, target_l, 0, 1).c,
+                        size);
+            }
+        }
+        else
+        {
+            Parallel_Reduce::reduce_double_allpool(kpar,
+                    GlobalV::NPROC_IN_POOL,
+                    occmat.mat(iat, target_l, 0, 0).c,
+                    size * 4);
+        }
+    }
+}
+
+void compute_pot_uterm_and_energy(const UnitCell& cell,
+                                  const int nspin,
+                                  const std::vector<double>& u_current,
+                                  const std::vector<int>& l_channel,
+                                  const std::vector<int>& uterm_mat_index,
+                                  const OccupationMatrix& occmat,
+                                  std::vector<std::complex<double>>& uterm_mat,
+                                  double& energy_u)
+{
+    energy_u = 0.0;
+    const double weight_eu = (nspin == 1) ? 1.0 : (nspin == 2) ? 0.5 : 0.25;
+    const double diag_coeff = (nspin == 4) ? 1.0 : 0.5;
+    // calculate pot_onsite and energy (occ_mat already reduced above)
+    for(int iat = 0; iat < cell.nat; iat++)
+    {
+        const int it = cell.iat2it[iat];
+        const int target_l = l_channel[it];
+        if(target_l == -1)
+        {
+            continue;
+        }
+        const int size = (2 * target_l + 1) * (2 * target_l + 1);
+
+        //update effective potential
+        const double u_value = u_current[it];
+        std::complex<double>* pot_onsite_iat = &(uterm_mat[uterm_mat_index[iat]]);
+        const int m_size = 2 * target_l + 1;
+
+        if(nspin == 4)
+        {
+            // pot_onsite is stored as 4 contiguous Pauli blocks per atom:
+            //   is=0: charge channel (identity), Hubbard U contributes the
+            //         diagonal term diag_coeff*delta(m1,m2)
+            //   is=1,2,3: spin channels (sigma_x/y/z), no U diagonal term
+            // The occupation matrix occ_mat[...][0][0].c packs all 4 blocks
+            // contiguously, each of size m_size*m_size.
+            energy_u += compute_pot_onsite_spinor(
+                pot_onsite_iat,
+                occmat.mat(iat, target_l, 0, 0).c,
+                u_value, diag_coeff, weight_eu, m_size);
+        }
+        else // nspin=1 or nspin=2
+        {
+            // spin-up channel
+            energy_u += compute_pot_onsite_scalar(
+                pot_onsite_iat,
+                occmat.mat(iat, target_l, 0, 0).c,
+                u_value, diag_coeff, weight_eu, m_size);
+            // spin-down channel for nspin=2
+            if(nspin == 2)
+            {
+                std::complex<double>* pot_onsite_iat1 = &(uterm_mat[uterm_mat.size()/2 + uterm_mat_index[iat]]);
+                energy_u += compute_pot_onsite_scalar(
+                    pot_onsite_iat1,
+                    occmat.mat(iat, target_l, 0, 1).c,
+                    u_value, diag_coeff, weight_eu, m_size);
+            }
+        }
+    }
+}
+
+} // namespace pw
