+++ source/source_pw/module_pwdft/dftu_pw_tools.h
@@ -1,5 +1,5 @@
-#ifndef DFTU_BASE_TOOLS_H
-#define DFTU_BASE_TOOLS_H
+#ifndef DFTU_PW_TOOLS_H
+#define DFTU_PW_TOOLS_H
 
 #include <complex>
 #include <vector>
@@ -7,16 +7,13 @@
 
 class UnitCell;
 class OccupationMatrix;
-class OccMatMixer;
-class Charge_Mixing;
 
-/// Free functions for DFT+U PW basis calculations.
+/// Pure (side-effect free) functions for DFT+U PW basis calculations.
 ///
-/// These functions are pure (no access to Plus_U_Base members) so they can be
-/// unit-tested directly by including this header. The member functions in
-/// dftu_base_occ.cpp call them after computing per-atom offsets and fetching
-/// the relevant member state (occ_mat, uterm_mat, u_current, etc.).
-namespace DFTU_BASE {
+/// These functions are stateless and can be unit-tested directly.
+/// The member functions in dftu_pw.cpp call them after computing
+/// per-atom offsets and fetching the relevant member state.
+namespace pw {
 
 /// transform pot_onsite from Pauli basis to spin basis (in-place, nspin==4 only).
 ///
@@ -136,46 +133,6 @@ void compute_pot_uterm_and_energy(const UnitCell& cell,
                                   std::vector<std::complex<double>>& uterm_mat,
                                   double& energy_u);
 
-/// accumulate occ_mat from psi for all k-points (per-device template).
-///
-/// Explicitly instantiated for DEVICE_CPU (and DEVICE_GPU when available)
-/// in dftu_base_occ.cpp.
-template <typename Device>
-void accumulate_occ_one_k(const void* psi_in,
-                          const ModuleBase::matrix& wg_in,
-                          const UnitCell& cell,
-                          const int* isk,
-                          const int nspin,
-                          const std::vector<int>& l_channel,
-                          OccupationMatrix& occmat);
+} // namespace pw
 
-/// calculate the local occupation number matrix for PW based wave functions.
-///
-/// This is the PW-basis entry point that:
-///   1. saves and zeroes the occupation matrix
-///   2. accumulates it from psi via accumulate_occ_one_k
-///   3. reduces across k-pools via reduce_occ_mat
-///   4. applies occupation-matrix mixing when enabled
-///   5. computes the effective potential and DFT+U energy
-///
-/// All state is passed explicitly so this function can be unit-tested
-/// without constructing a Plus_U_Base object.
-void cal_occ_pw(const void* psi_in,
-                const ModuleBase::matrix& wg_in,
-                const UnitCell& cell,
-                Charge_Mixing* p_chgmix,
-                const int* isk,
-                const int kpar,
-                const int nspin,
-                const std::string& device,
-                const std::vector<int>& l_channel,
-                const std::vector<double>& u_current,
-                const std::vector<int>& uterm_mat_index,
-                OccupationMatrix& occmat,
-                OccMatMixer* occ_mixer,
-                std::vector<std::complex<double>>& uterm_mat,
-                double& energy_u);
-
-} // namespace DFTU_BASE
-
-#endif
+#endif // DFTU_PW_TOOLS_H
