+++ source/source_pw/module_pwdft/setup_dftu_pw.cpp
@@ -1,7 +1,7 @@
 #include "source_pw/module_pwdft/setup_dftu_pw.h"
 #include "source_pw/module_pwdft/dftu_base.h" // mohan add 2025-11-06
 #include "source_pw/module_pwdft/dftu_base_io.h" // mohan add 2025-11-08
-#include "source_pw/module_pwdft/dftu_base_tools.h"
+#include "source_pw/module_pwdft/dftu_pw.h"
 #include "source_io/module_parameter/parameter.h"
 
 namespace DFTU_BASE
