+++ source/source_pw/module_pwdft/test/dftu_base_test.cpp
@@ -16,6 +16,7 @@
 #include "gtest/gtest.h"
 
 #include <vector>
+#include <numeric>
 
 class DFTUBaseTest : public testing::Test
 {
@@ -118,4 +119,262 @@ TEST_F(DFTUBaseTest, InitBaseYukawaFalseThenTrueCreatesObject)
 
     init_dftu(dftu, true);
     EXPECT_TRUE(dftu.use_yukawa());
+}
+
+// =====================================================================
+// uterm_mat_index calculation
+//
+// nspin=1: offset = sum(tlp1^2), total = sum(all tlp1^2)
+// nspin=2: same per-spin-channel, then pot_index *= 2 (split layout)
+// nspin=4: offset = sum((tlp1*npol)^2), each atom = 4*tlp1^2
+// =====================================================================
+
+class EffPotIndexTest : public ::testing::Test
+{
+  protected:
+    struct AtomSpec { int l; int na; }; // correlated orbital l, number of atoms
+    std::vector<int> uterm_mat_index;
+    int pot_index;
+
+    void compute_indices(const std::vector<AtomSpec>& atoms, int nspin)
+    {
+        pot_index = 0;
+        uterm_mat_index.resize(atoms.size());
+
+        for (size_t i = 0; i < atoms.size(); i++)
+        {
+            int tlp1 = 2 * atoms[i].l + 1;
+            int tlp1_npol = tlp1 * (nspin == 4 ? 2 : 1);
+
+            if (nspin == 4)
+            {
+                uterm_mat_index[i] = pot_index;
+                pot_index += tlp1_npol * tlp1_npol;
+            }
+            else
+            {
+                uterm_mat_index[i] = pot_index;
+                pot_index += tlp1 * tlp1;
+            }
+        }
+
+        if (nspin == 2)
+            pot_index *= 2;
+    }
+};
+
+TEST_F(EffPotIndexTest, Nspin1_MixedOrbitals)
+{
+    // 3 atoms: p(l=1), d(l=2), p(l=1)
+    std::vector<AtomSpec> atoms = {{1, 1}, {2, 1}, {1, 1}};
+    compute_indices(atoms, 1);
+
+    // p: 9, d: 25, p: 9
+    EXPECT_EQ(uterm_mat_index[0], 0);
+    EXPECT_EQ(uterm_mat_index[1], 9);
+    EXPECT_EQ(uterm_mat_index[2], 34);
+    EXPECT_EQ(pot_index, 43); // 9 + 25 + 9
+}
+
+TEST_F(EffPotIndexTest, Nspin2and4_SplitAndPauli)
+{
+    // nspin=2: 2 d-atoms, split layout [up | dn]
+    std::vector<AtomSpec> atoms2 = {{2, 1}, {2, 1}};
+    compute_indices(atoms2, 2);
+    EXPECT_EQ(uterm_mat_index[0], 0);
+    EXPECT_EQ(uterm_mat_index[1], 25);
+    EXPECT_EQ(pot_index, 100); // (25 + 25) * 2
+
+    // nspin=4: d + p atoms, Pauli blocks
+    std::vector<AtomSpec> atoms4 = {{2, 1}, {1, 1}};
+    compute_indices(atoms4, 4);
+    EXPECT_EQ(uterm_mat_index[0], 0);    // d: (5*2)^2 = 100
+    EXPECT_EQ(uterm_mat_index[1], 100);  // p: (3*2)^2 = 36
+    EXPECT_EQ(pot_index, 136);
+}
+
+// =====================================================================
+// copy_occ_mat <-> set_occ_mat roundtrip
+//
+// Tests the bidirectional conversion between nested occ_mat matrix
+// and flat uom_array/uom_save arrays for all 3 nspin modes.
+// =====================================================================
+
+struct Matrix2D {
+    int nr, nc;
+    std::vector<double> data;
+    Matrix2D() : nr(0), nc(0), data() {}
+    Matrix2D(int r, int c) : nr(r), nc(c), data(r * c, 0.0) {}
+    double& operator()(int i, int j) { return data[i * nc + j]; }
+    const double& operator()(int i, int j) const { return data[i * nc + j]; }
+};
+
+static void copy_occ_mat_to_flat(
+    const std::vector<Matrix2D>& occ_mat_up,
+    const std::vector<Matrix2D>& occ_mat_dn,
+    std::vector<double>& uom_save,
+    const std::vector<int>& uterm_mat_index,
+    int nspin)
+{
+    if (nspin == 4)
+    {
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+                uom_save[uterm_mat_index[iat] + mm] = occ_mat_up[iat].data[mm];
+        }
+    }
+    else if (nspin == 2) // split layout: [up | dn]
+    {
+        int half_size = uom_save.size() / 2;
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+            {
+                uom_save[uterm_mat_index[iat] + mm] = occ_mat_up[iat].data[mm];
+                uom_save[half_size + uterm_mat_index[iat] + mm] = occ_mat_dn[iat].data[mm];
+            }
+        }
+    }
+    else // nspin=1: single spin channel
+    {
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+                uom_save[uterm_mat_index[iat] + mm] = occ_mat_up[iat].data[mm];
+        }
+    }
+}
+
+static void set_occ_mat_from_flat(
+    const std::vector<double>& uom_array,
+    std::vector<Matrix2D>& occ_mat_up,
+    std::vector<Matrix2D>& occ_mat_dn,
+    const std::vector<int>& uterm_mat_index,
+    int nspin)
+{
+    if (nspin == 4)
+    {
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+                occ_mat_up[iat].data[mm] = uom_array[uterm_mat_index[iat] + mm];
+        }
+    }
+    else if (nspin == 2)
+    {
+        int half_size = uom_array.size() / 2;
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+            {
+                occ_mat_up[iat].data[mm] = uom_array[uterm_mat_index[iat] + mm];
+                occ_mat_dn[iat].data[mm] = uom_array[half_size + uterm_mat_index[iat] + mm];
+            }
+        }
+    }
+    else // nspin=1
+    {
+        for (size_t iat = 0; iat < occ_mat_up.size(); iat++)
+        {
+            int size = occ_mat_up[iat].nr * occ_mat_up[iat].nc;
+            for (int mm = 0; mm < size; mm++)
+                occ_mat_up[iat].data[mm] = uom_array[uterm_mat_index[iat] + mm];
+        }
+    }
+}
+
+class OccMatRoundtripTest : public ::testing::Test
+{
+  protected:
+    void SetUp() override {}
+};
+
+TEST_F(OccMatRoundtripTest, Nspin1and2_SingleAndSplitLayout)
+{
+    // nspin=1: single atom d-orbital roundtrip
+    const int l = 2;
+    const int size = (2 * l + 1) * (2 * l + 1); // 25
+
+    std::vector<Matrix2D> occ_mat_up(1, Matrix2D(2 * l + 1, 2 * l + 1));
+    std::vector<Matrix2D> occ_mat_dn(1, Matrix2D(2 * l + 1, 2 * l + 1));
+    for (int i = 0; i < size; i++)
+        occ_mat_up[0].data[i] = static_cast<double>(i + 1);
+
+    std::vector<int> uterm_mat_index = {0};
+    std::vector<double> uom_save(size, 0.0);
+    copy_occ_mat_to_flat(occ_mat_up, occ_mat_dn, uom_save, uterm_mat_index, 1);
+    set_occ_mat_from_flat(uom_save, occ_mat_up, occ_mat_dn, uterm_mat_index, 1);
+    for (int i = 0; i < size; i++)
+        EXPECT_DOUBLE_EQ(occ_mat_up[0].data[i], static_cast<double>(i + 1));
+
+    // nspin=2: split layout [up | dn] with distinct values
+    const int total = size * 2;
+    for (int i = 0; i < size; i++)
+    {
+        occ_mat_up[0].data[i] = static_cast<double>(i + 1);
+        occ_mat_dn[0].data[i] = static_cast<double>(i + 100);
+    }
+    uom_save.assign(total, 0.0);
+    copy_occ_mat_to_flat(occ_mat_up, occ_mat_dn, uom_save, uterm_mat_index, 2);
+    // Verify split layout
+    for (int i = 0; i < size; i++)
+    {
+        EXPECT_DOUBLE_EQ(uom_save[i], static_cast<double>(i + 1));
+        EXPECT_DOUBLE_EQ(uom_save[size + i], static_cast<double>(i + 100));
+    }
+    set_occ_mat_from_flat(uom_save, occ_mat_up, occ_mat_dn, uterm_mat_index, 2);
+    for (int i = 0; i < size; i++)
+    {
+        EXPECT_DOUBLE_EQ(occ_mat_up[0].data[i], static_cast<double>(i + 1));
+        EXPECT_DOUBLE_EQ(occ_mat_dn[0].data[i], static_cast<double>(i + 100));
+    }
+}
+
+TEST_F(OccMatRoundtripTest, Nspin4_PauliBlocks)
+{
+    // 2 atoms: d(l=2), p(l=1)
+    struct AtomSpec { int l; };
+    std::vector<AtomSpec> specs = {{2}, {1}};
+    int npol = 2;
+
+    std::vector<int> sizes;
+    for (auto& s : specs)
+    {
+        int tlp1 = 2 * s.l + 1;
+        sizes.push_back((tlp1 * npol) * (tlp1 * npol));
+    }
+    int total = std::accumulate(sizes.begin(), sizes.end(), 0);
+
+    std::vector<int> uterm_mat_index(specs.size());
+    int offset = 0;
+    for (size_t i = 0; i < specs.size(); i++)
+    {
+        uterm_mat_index[i] = offset;
+        offset += sizes[i];
+    }
+
+    std::vector<Matrix2D> occ_mat(specs.size());
+    for (size_t i = 0; i < specs.size(); i++)
+    {
+        int dim = (2 * specs[i].l + 1) * npol;
+        occ_mat[i] = Matrix2D(dim, dim);
+        for (int j = 0; j < sizes[i]; j++)
+            occ_mat[i].data[j] = static_cast<double>(i * 1000 + j + 1);
+    }
+
+    std::vector<double> uom_array(total, 0.0);
+    std::vector<Matrix2D> occ_mat_dn(specs.size()); // unused for nspin=4
+
+    copy_occ_mat_to_flat(occ_mat, occ_mat_dn, uom_array, uterm_mat_index, 4);
+    set_occ_mat_from_flat(uom_array, occ_mat, occ_mat_dn, uterm_mat_index, 4);
+
+    for (size_t i = 0; i < specs.size(); i++)
+        for (int j = 0; j < sizes[i]; j++)
+            EXPECT_DOUBLE_EQ(occ_mat[i].data[j], static_cast<double>(i * 1000 + j + 1));
 }
